+++ src/demo/polygon_visibility.h
@@ -7,7 +7,7 @@ class polygon_visibility:
 {
 public:
     void preprocessing(const std::vector<float>& points) override;
-    std::vector<float> get_visibility(double x, double y) override;
+    std::vector<float> get_visibility(float x, float y) override;
 private:
     
     std::vector<triangle*> m_triangles;
